refactor(turn-deadline): 产品路径不装闸,benchmark 路径换成 wall-clock(ADR-24)
轨迹(dh 用户安排行程)显示原「步数预算」是评测闸,误装到产品路径
导致模型在 EXHAUSTED 后继续派发工具、token 截断、用户拿不到行程。
ts/src/tool-budget.ts → ts/src/turn-deadline.ts,语义换为 wall-clock
(默认 60s 软阈 / 120s 硬阈,env 可覆盖);产品路径 install() 默认
不调用,benchmark opt-in 才装;同步收紧工具 schema(同一 step 内
不会再回环);run-all §45 改名 + 加 enabled:false 产品路径不变量断言。
