+++ README.zh-CN.md
@@ -174,7 +174,7 @@ node scripts/build-dist.mjs                       # 构建 JS runtime
 - **账号会话检索** —— 你本人登录态查携程机票;观测轮次中所有可评分 hit 全过、ReadGuard 零写,非 hit 保持显式 `miss` 记录——不作超出此口径的实时可售声明
 - **扩展按需装** —— `[GoTry Session Bridge](https://chromewebstore.google.com/detail/gotry-session-bridge/oeajpiccmonococjcegddlooeeohlbgd)` 由 dsh 宿主 UI 在账号会话工具首次需要时以可点链接给出(Chrome 商店一键装 + 自动更新);gotry 这边不跑 setup wizard、不开 chrome://extensions、不动剪贴板
 - **记忆与触达** —— 动机画像 / 愿望池 / 同行人 / 旅行时间线;英文输出一键切换(`GOTRY_LOCALE=en`)
-- **有界的工具循环** —— 16 次派发后注入软收敛,18 次后结构化 `TOOL_BUDGET_EXHAUSTED` 拒绝;经打包消费者安装的 E2E 在 CI 里实测
+- **有界的人工回合** —— 产品路径不装任何「工具调用次数」闸;wall-clock 截止(默认 60s 软阈 / 120s 硬阈,可用 `GOTRY_TURN_DEADLINE_SOFT_MS` / `GOTRY_TURN_DEADLINE_HARD_MS` 覆盖)仅在 benchmark opt-in 路径启用,超硬阈后结构化 `TURN_DEADLINE_EXHAUSTED` 拒绝;经打包消费者安装的 E2E 在 CI 里实测
 
 **已知限制**(诚实清单):
