+++ docs/roadmap.md
@@ -46,7 +46,7 @@ Evaluation Phase 0 foundation boundary: contracts/registry/validators/unmatched 
 ### Agent evaluation 反馈闭环(Discussion #78)
 
 - ChinaTravel grounding-v3 冻结 canary 的首例可评分终态通过,第二例因 planner 重复工具循环超过 300s；当前只允许报告 `attempted=2/scored=1/timeout=1`,不得外推 5-query aggregate。
-- Round 1 聚焦 GoTry 自身:模型每轮第 16 次真实工具派发注入收敛上下文,第 18 次是最后一个 body；同一步第 19 次起返回 `TOOL_BUDGET_EXHAUSTED`,该批 `step/end` 后下一步 native 请求抑制继承工具 schema。run-all §45 已覆盖 Cordis integration/并行/turn+session 生命周期及真实 npm-mode dist + dsh headless 离线 E2E(18 body、1 structured refusal、下一请求 text-only、非空 final)；CI 以当前 SHA tarball 的隔离 pnpm consumer 入口重放同一链路,不借 root 开发树。
+- Round 1 聚焦 GoTry 自身:模型每轮第 16 次真实工具派发注入收敛上下文,第 18 次是最后一个 body；同一步第 19 次起返回 `TOOL_BUDGET_EXHAUSTED`,该批 `step/end` 后下一步 native 请求抑制继承工具 schema。run-all §45 已覆盖 Cordis integration/并行/turn+session 生命周期及真实 npm-mode dist + dsh headless 离线 E2E(18 body、1 structured refusal、下一请求 text-only、非空 final)；CI 以当前 SHA tarball 的隔离 pnpm consumer 入口重放同一链路,不借 root 开发树。**Round 1 后续(ADR-24)**:真实轨迹(用户安排行程)显示「步数预算」对产品用户体验是负向的——`TOOL_BUDGET_EXHAUSTED` 作为 tool result 回灌后模型在同一 step 内继续派发 3+ 次工具调用、token 截断、用户拿不到答复。闸语义从「步数」换为「wall-clock」(默认 60s/120s,可由 env 覆盖);产品路径(`apply()` 默认分支)不再装闸,benchmark opt-in 路径才装——产品用户不再被评测闸误伤。run-all §45 改名 `agent-planning-turn-deadline-{tests,e2e}.ts`,断言集合加上 `enabled: false` 时 `ctx.on` 不被订阅的产品路径不变量。
 - 外部验收仍是 TODO:Round 1 treatment-attested installed GoTry 的 exact DeepSeek 为 environment unavailable/schema-invalid 0，GLM 为 300 秒 timeout。Round 2 的唯一冻结 treatment 只算 diagnostic-only：agent 没有把 CLI/shell 指令转成唯一 native bridge call，终态也没有 tagged JSON。Round 3 修复该通用 conformance 缺口后，新冻结 treatment 的 provider preflight 通过且 runner 只派发一次，但 planner/runner 均 exit 1，3.796 秒内停止，parent 释放 0 字节，evaluator 未进入，official scores 仍全为 null。Round 4 treatment（SHA `5ebddb2`）primary preflight=pass，但 planner/runner 均在 30.968s 以 exit 1 结束，释放 0 字节，evaluator 未进入，official scores 全为 null；产品 gate 是 Node v24.20.0，而 treatment 使用 v26.3.0，因此仅 diagnostic-only、无 uplift。GitHub Node 22/24 §48 另暴露 source default-off 30s lifecycle hang。Round 5 仅限移除 timer/keepalive preload、将锁定的 DSH runtime closure 设为 alpha.3、保留源码普通运行的 `ts/dsh-runtime/gotry-state/` 状态连续性、让 benchmark/package 用调用目录隔离，并增加 benchmark-only 结构化诊断 pipe（allowlisted redacted reason codes），stdout 继续 fail-closed；其 frozen treatment（代码 SHA `752e54c`）在 140.715 秒后以 `child_nonzero_exit`、0 terminal bytes、evaluator/official scores null 停止，仅 diagnostic-only，后续纯 lock-consistency 提交不改写该 UID 归属。Round 6 frozen treatment 已把失败稳定归因为 `child_runtime_error`，但仍未产生可评分终态；七指标与 overall 仍待独立证据，不得声称 external benchmark closure。
 - **Phase 1 bridge 当前态**:Round 2 已有 default-off owner-local bridge contract，固定 executable/cwd/argv_prefix/allowlist 与 timeout/output caps，并可按工具配置 `allowed_output_keys` 正向 visible-output 键合同，未声明键 fail-closed 且不反射。opt-in 仅允许 cold-start；每个 benchmark agent 强制 native、原始 bridge definition-only，最终 assembly 与 downstream pre-step 双闸拒绝额外 schema/同名 shadow/PTC 漂移。Round 3 增加 provider-neutral conformance：system contract 把 CLI/shell/Python 操作映射成 `query.action=call`，只有 allowed call 的配对成功 result 后才接受 owner-local `terminal_output` 指定的单一 tagged JSON object；缺 call 或坏终态最多纠偏一次，格式纠偏禁止重复派发 bridge，headless parent 用同一 parser 二次验收 stdout。Round 4 在 config 验证后、任何 optional host plugin 解析前，把 benchmark patch 的顶层 insert 投影为唯一 `gotry-tools`；inline/reordered/未来非 GoTry insert 全部裁掉，`gotry-tools` 或 config-path 注入锚点缺失/重复均在 spawn/relay 前 fail-closed，default-off 组合不变。agent disposal 释放 scoped effect；插件先卸载时 live agent 保持 fail-closed quarantine 到 agent/process disposal，HMR 不热挂。config 要求 writes forbidden/network denied，但真实 enforcement 由 owner/host OS 负责，synthetic E2E 不证明 OS sandbox。bridge 不引入 Python/benchmark runtime 产品依赖。Round 4 treatment（SHA `5ebddb2`）primary preflight=pass，但 planner/runner 均在 30.968s 以 exit 1 结束，释放 0 字节，evaluator 未进入，official scores 全为 null；产品 gate 使用 Node v24.20.0，而 treatment 使用 v26.3.0，因此仅 diagnostic-only、无 uplift。GitHub Node 22/24 §48 另暴露 source default-off 30s lifecycle hang。Round 5 仅限移除 timer/keepalive preload、将锁定的 DSH runtime closure 设为 alpha.3、保留源码普通运行的 `ts/dsh-runtime/gotry-state/` 状态连续性、让 benchmark/package 用调用目录隔离，并增加 benchmark-only 结构化诊断 pipe（allowlisted redacted reason codes），stdout 继续 fail-closed；其 frozen treatment（代码 SHA `752e54c`）在 140.715 秒后以 `child_nonzero_exit`、0 terminal bytes、evaluator/official scores null 停止，仅 diagnostic-only，后续纯 lock-consistency 提交不改写该 UID 归属。
 - **Round 5 runtime resolution**:source checkout 与 clean package 都解析同一组 216 个精确直接依赖的 DSH `0.1.2-alpha.3` closure；manifest、package-lock 与 root pnpm importer 的 DSH 名称集合必须同为 216，publish preverify 对漏钉、混版或 range 声明 fail-closed；source 普通运行的 dsh cwd 保持 `ts/dsh-runtime/`，状态继续落 `ts/dsh-runtime/gotry-state/`，benchmark opt-in 与 npm package 使用调用目录隔离；legacy vendored alpha.1 仅作非 benchmark 解析兼容，不承诺可运行。benchmark 在 spawn 前拒绝实际版本漂移；Node 下界统一为 22.15，并在 DSH import/spawn 前 fail-closed。legacy vendor 的整体升级留在独立 PR，不混入本轮。
