+++ docs/stage1-top-down-design.md
@@ -7,7 +7,7 @@
 - Phase 1 environment bridge 当前态:default-off owner-local config,固定 executable/cwd/argv_prefix/allowlist 与 timeout/output caps，可选 per-tool `allowed_output_keys` 正向 visible-output 键合同对未声明键 fail-closed 且不反射。benchmark opt-in 是 cold-start + headless one-shot 进程边界，不能热挂 live agent；每次请求仅允许 native、原始 bridge definition-only surface，最终 assembly 与 downstream pre-step 双闸 fail-closed。Round 3 conformance 要求至少一次 allowed `query.action=call` 及其配对成功 result，随后只接受 owner-local `terminal_output` 指定的单一 tagged JSON object；缺 call 或坏终态最多一次固定纠偏，格式纠偏禁止重复派发 bridge，parent stdout 再过同一 parser。Round 4 在 optional plugin 解析前把 benchmark 顶层 insert 投影为唯一 `gotry-tools`，未知/未来 sibling 安全裁掉；allowlist 与 config-path anchor 缺失/重复均在 dsh spawn/relay 前 fail-closed，default-off 组合不变。局部安装回滚与 agent disposal 属 S5 双层 effect 合同；插件先卸载时 live agent 保持 fail-closed quarantine 到 agent/process disposal，HMR 不热挂。config 要求 writes forbidden/network denied，但实际 enforcement 由 owner/host OS 负责，synthetic E2E 不证明 OS sandbox。Round 4 treatment（SHA `5ebddb2`）primary preflight=pass，但 planner/runner 均在 30.968s 以 exit 1 结束，释放 0 字节，evaluator 未进入，official scores 全为 null；产品 gate 使用 Node v24.20.0，而 treatment 使用 v26.3.0，因此仅 diagnostic-only、无 uplift。GitHub Node 22/24 §48 另暴露 source default-off 30s lifecycle hang。Round 5 仅限移除 timer/keepalive preload、将锁定的 DSH runtime closure 设为 alpha.3、保留源码普通运行的 `ts/dsh-runtime/gotry-state/` 状态连续性、让 benchmark/package 用调用目录隔离，并增加 benchmark-only 结构化诊断 pipe（allowlisted redacted reason codes），stdout 继续 fail-closed；其 frozen treatment（代码 SHA `752e54c`）在 140.715 秒后以 `child_nonzero_exit`、0 terminal bytes、evaluator/official scores null 停止，仅 diagnostic-only，后续纯 lock-consistency 提交不改写该 UID 归属。Phase 0 foundation 的 deterministic planning/admission 边界不变。
 - Round 5 runtime resolution:source checkout 与 clean package 都解析同一组 216 个精确直接依赖的 DSH `0.1.2-alpha.3` closure；manifest、package-lock 与 root pnpm importer 的 DSH 名称集合必须同为 216，publish preverify 对漏钉、混版或 range 声明 fail-closed。source 普通运行的 dsh cwd 保持 `ts/dsh-runtime/`，状态继续落 `ts/dsh-runtime/gotry-state/`，benchmark opt-in 与 npm package 使用调用目录隔离；legacy vendored alpha.1 只作非 benchmark 解析兼容，不承诺可运行。benchmark 在 spawn 前拒绝实际 DSH 版本漂移；Node 下界统一为 22.15，并在 DSH import/spawn 前稳定拒绝旧版本。
 - Round 6 terminal diagnostics:benchmark conformance 只观察最终 `turn/end`，按 allowlisted code/有限 status 映射 coarse enum；所有自由文本与 raw stderr 均不进入控制面。per-session arbiter 只在最终终态单写，bridge/conformance 专项原因优先，恢复成功的 model retry 零失败输出。代码 SHA `c61600b` 的 ChinaTravel frozen treatment（`..._00001`，`deepseek-v4-flash`）在 49.546 秒后稳定输出 `child_runtime_error`，但仍为 0 terminal bytes、evaluator/official scores null；leakage 与本地凭证/端点扫描均为 0。该切片不改变重试、prompt、工具和评分逻辑，不产生 uplift。
-- Round 7 minimal kernel:代码 SHA `edb9392896625adbb48abae4a2ecf968dbfc0349` 的 benchmark opt-in 仅保留 tool budget、model override、唯一 bridge、isolation/conformance；产品 prompt variables/process guards/consent/普通工具均不装，默认路径不变。system-prompt 投影为稳定 task-agnostic persona，root patch 只接受 canonical `insert` 与 `system-prompt` 各一，其他 root item/变体 fail-closed。ChinaTravel frozen treatment UID `e20241028160248698752`（`easy`，`deepseek-v4-flash`）preflight pass、未回退，80.463s 后 runner exit 1、terminal 0 bytes/invalid，evaluator 未进入、official null、不可计分；白名单归因为 `child_bridge_runner_failed`，不产生 uplift 或 external benchmark closure。下一轮问题聚焦 generic bridge tool schema 与可恢复 domain-error contract。
+- Round 7 minimal kernel:代码 SHA `edb9392896625adbb48abae4a2ecf968dbfc0349` 的 benchmark opt-in 仅保留 turn deadline(产品路径默认不装;benchmark opt-in 装 wall-clock 闸)、model override、唯一 bridge、isolation/conformance；产品 prompt variables/process guards/consent/普通工具均不装，默认路径不变。system-prompt 投影为稳定 task-agnostic persona，root patch 只接受 canonical `insert` 与 `system-prompt` 各一，其他 root item/变体 fail-closed。ChinaTravel frozen treatment UID `e20241028160248698752`（`easy`，`deepseek-v4-flash`）preflight pass、未回退，80.463s 后 runner exit 1、terminal 0 bytes/invalid，evaluator 未进入、official null、不可计分；白名单归因为 `child_bridge_runner_failed`，不产生 uplift 或 external benchmark closure。下一轮问题聚焦 generic bridge tool schema 与可恢复 domain-error contract。
 - ** 2026-08-27 起,①③ 两环的时间语义由 `ts/src/time-anchor.ts` 锚点卡供给(ADR-12),本文①中「2026 年历」的硬编码表述以锚点层为准。
 - 2026-08-28 起,S5 工单持久化升级为账本 durable 形态(ADR-15:workflow_runs/steps 权威 + json/md 视图),本文工单文件表述以 `state-ledger.ts` 为准；`gotry_async_terminal.v1` 将 4/4 映射为 `succeeded`/ledger `settled`/exit 0，将非 4/4 映射为 `failed`/ledger `failed`/exit 2，终态复诵保持同一结果与退出码且零重算。2026-08-29 起,求解运行时收敛 `z3-shared.ts`(单一 WASM 实例+会话级互斥,run-all §30 并发回归闸)——
 - Z3 WASM race 已知限制清偿。2026-08-29 同批:④中 dated 段经 `realtime-pricing.ts` 实时价覆写(env 闸默认关,证据 `[实时API:flyai@ts]`),run-all §31。同批:i18n catalog 接缝(`i18n.ts`,`GOTRY_LOCALE` 默认 zh-CN,金标准逐字节不变),求解确定性面英文可用,run-all §32。
